chat/requirements.txt
pydantic
fastapi[standard]
sqlalchemy
pyjwt
requests
psycopg2-binary